[ui] Extract search results into separate component SearchResultTable
